secure booking error

+++ HotelBookingSystem/src/main/java/org/swp391/hotelbookingsystem/service/BookingService.java
@@ -30,8 +30,10 @@ public class BookingService {
         bookingRepo.approveBookingUnit(id);
     }
 
-    public void deletePendingBooking(int id){
-        bookingRepo.deletePendingBooking(id);
+    public void deletePendingBooking(int id, int userId){
+        if(bookingRepo.idPending(id, userId)){
+            bookingRepo.deletePendingBooking(id);
+        }
     }
 
     public int saveBooking(Booking booking){
